requirements.txt
botocore==1.15.45
boto3==1.12.45
certifi==2019.11.28
chardet==3.0.4
Django==2.1.5
dj-database-url==0.5.0
django-heroku==0.3.1
django-crispy-forms==1.9.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
Pillow==7.1.1
psycopg2-binary==2.8.4
pyp==2.12
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
whitenoise==5.0.1
